Fixing custom actions
