+++ installer/ca.wxs
@@ -1,11 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<Wix xmlns="http://schemas.microsoft.com/wix/2006/wi">
-	<Fragment>
-    <CustomAction Id='IsRepositoryRunningAction' BinaryKey='CABin' DllEntry='IsRepositoryRunning' Execute='immediate'
+<Wix xmlns="http://schemas.microsoft.com/wix/2006/wi" xmlns:util="http://schemas.microsoft.com/wix/UtilExtension">
+  <Fragment>
+    <CustomAction Id='IsRepositoryRunningAction' BinaryKey='CABin' DllEntry='IsRepositoryRunning' Impersonate='no' Execute='deferred'
                   Return='check'/>
-    <CustomAction Id='ImportAppAction' BinaryKey='CABin' DllEntry='ImportApp' Execute='immediate'
+    <CustomAction Id='ImportAppAction' BinaryKey='CABin' DllEntry='ImportApp' Impersonate='no' Execute='deferred'
+                  Return='check'/>
+    <CustomAction Id='ImportExtensionsAction' BinaryKey='CABin' DllEntry='ImportExtensions' Impersonate='no' Execute='deferred'
                   Return='check'/>
 
-    <Binary Id='CABin' SourceFile='./CustomActions/bin/x64/Debug/CustomActions.dll'/>
+    <Binary Id='CABin' SourceFile='./CustomActions/bin/x64/Debug/CustomActions.CA.dll'/>
 	</Fragment>
 </Wix>
