+++ installer/installer.sln
@@ -5,30 +5,37 @@ VisualStudioVersion = 15.0.26430.16
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "installer", "installer.wixproj", "{602C8A92-6222-4B60-A603-8ECDA8BB587C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomActions", "CustomActions\CustomActions.csproj", "{B46C290D-3FF4-4CB1-B076-F30548A91164}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomActions", "CustomActions\CustomActions.csproj", "{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
 		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Debug|x64.ActiveCfg = Debug|x64
 		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Debug|x64.Build.0 = Debug|x64
+		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Debug|x86.ActiveCfg = Debug|x86
+		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Debug|x86.Build.0 = Debug|x86
 		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Release|Any CPU.ActiveCfg = Release|x86
 		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Release|x64.ActiveCfg = Release|x64
 		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Release|x64.Build.0 = Release|x64
-		{B46C290D-3FF4-4CB1-B076-F30548A91164}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B46C290D-3FF4-4CB1-B076-F30548A91164}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B46C290D-3FF4-4CB1-B076-F30548A91164}.Debug|x64.ActiveCfg = Debug|x64
-		{B46C290D-3FF4-4CB1-B076-F30548A91164}.Debug|x64.Build.0 = Debug|x64
-		{B46C290D-3FF4-4CB1-B076-F30548A91164}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B46C290D-3FF4-4CB1-B076-F30548A91164}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B46C290D-3FF4-4CB1-B076-F30548A91164}.Release|x64.ActiveCfg = Release|Any CPU
-		{B46C290D-3FF4-4CB1-B076-F30548A91164}.Release|x64.Build.0 = Release|Any CPU
+		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Release|x86.ActiveCfg = Release|x86
+		{602C8A92-6222-4B60-A603-8ECDA8BB587C}.Release|x86.Build.0 = Release|x86
+		{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}.Debug|x64.ActiveCfg = Debug|x64
+		{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}.Debug|x64.Build.0 = Debug|x64
+		{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}.Debug|x86.ActiveCfg = Debug|x86
+		{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}.Debug|x86.Build.0 = Debug|x86
+		{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}.Release|Any CPU.ActiveCfg = Release|x86
+		{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}.Release|x64.ActiveCfg = Release|x86
+		{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}.Release|x86.ActiveCfg = Release|x86
+		{DC4F450C-71F1-4748-88E8-1FADA6F38B0A}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
